src/backend/requirements.txt
fastapi==0.123.7
uvicorn[standard]==0.38.0
langchain_classic==1.0.0
langchain_community==0.4.1
langchain_core==1.1.0
langchain_google_genai==3.2.0
langchain_huggingface==1.1.0
langchain_text_splitters==1.0.0
pydantic==2.12.5
python-dotenv==1.2.1
pgvector==0.4.2
psycopg2-binary==2.9.11
sentence-transformers==5.1.2
# pysrt==1.1.2
